[Refactor] move some functions to test utils
